Create README.md
Built Power BI dashboard to visualize the trends of the EV Vehicles in the market. Insights shows that there is the huge potential growth opportunities of the Battery Electric Vehicle. Power BI for data visualization and interactive dashboards. Kaggle website is used to load the dataset into the excel file and Data cleaning, Data processing all are done in MS EXCEL by using EXCEL Methods.